requirements.txt
lightning_bolts>=0.5.0
numpy>=1.20.3
Pillow>=9.2.0
pytorch_lightning>=1.5.2
ruamel.base>=1.0.0
torch>=1.11.0
torchmetrics>=0.6.0
torchvision>=0.12.0
wget>=3.2
